Extract the abstract concern out of Pin::Base
Adding abstract state to Pin::Base pushed it past Metrics/ClassLength at
509 lines, which failed the rubocop job.

Pin::Abstractable now holds abstract? and abstract_note, following the
mixin-dependency convention Equality already uses. Documenting gains an
abstract_note returning nil that Abstractable overrides, so the note is
rendered without a super call: a module has no statically resolvable
parent, and overriding #documentation from the mixin reported two
Unresolved call to super problems. It also keeps Documenting usable on its
own, which documenting_spec relies on.

Also drops a doc comment orphaned above authority_over when
choose_priority was replaced.
