+++ lib/solargraph/pin.rb
@@ -10,6 +10,7 @@ module Solargraph
   module Pin
     autoload :Common,           'solargraph/pin/common'
     autoload :Conversions,      'solargraph/pin/conversions'
+    autoload :Abstractable,     'solargraph/pin/abstractable'
     autoload :Base,             'solargraph/pin/base'
     autoload :Method,           'solargraph/pin/method'
     autoload :Signature,        'solargraph/pin/signature'
